+++ watershed-android/watershed/app/src/main/java/com/blueprint/watershed/Tasks/TaskList/TaskListAbstractFragment.java
@@ -14,8 +14,10 @@ import android.widget.ExpandableListView;
 
 import com.android.volley.Response;
 import com.blueprint.watershed.Activities.MainActivity;
+import com.blueprint.watershed.MiniSites.MiniSite;
 import com.blueprint.watershed.Networking.NetworkManager;
 import com.blueprint.watershed.Networking.Tasks.TaskListRequest;
+import com.blueprint.watershed.Networking.Users.UserMiniSitesRequest;
 import com.blueprint.watershed.R;
 import com.blueprint.watershed.Tasks.CreateTaskFragment;
 import com.blueprint.watershed.Tasks.Task;
@@ -42,6 +44,8 @@ public abstract class TaskListAbstractFragment extends ListFragment {
     protected List<String> mTaskListHeaders;
     protected TaskAdapter mTaskAdapter;
 
+    protected ArrayList<MiniSite> mUserMiniSiteList;
+
     protected FloatingActionButton mCreateTask;
     protected ExpandableListView mListView;
     protected SwipeRefreshLayout mNoTasks;
@@ -55,6 +59,7 @@ public abstract class TaskListAbstractFragment extends ListFragment {
         mNetworkManager = NetworkManager.getInstance(mParentActivity);
         mTaskList = new HashMap<String, List<Task>>();
         mTaskListHeaders = new ArrayList<String>();
+        mUserMiniSiteList = new ArrayList<MiniSite>();
     }
 
     @Override
@@ -62,6 +67,7 @@ public abstract class TaskListAbstractFragment extends ListFragment {
         super.onCreateView(inflater, container, savedInstanceState);
         View finalView = inflater.inflate(R.layout.fragment_task_list, container, false);
         initializeViews(finalView);
+        getMiniSiteRequest();
         getTasksRequest();
         return finalView;
     }
@@ -150,6 +156,21 @@ public abstract class TaskListAbstractFragment extends ListFragment {
         mNetworkManager.getRequestQueue().add(taskListRequest);
     }
 
+    /**
+     * Gets all minisites that a user has subscribed to.
+     */
+    protected void getMiniSiteRequest() {
+        UserMiniSitesRequest miniSitesRequest = new UserMiniSitesRequest(mParentActivity,
+                new HashMap<String, JSONObject>(),
+                new Response.Listener<ArrayList<MiniSite>>() {
+                    @Override
+                    public void onResponse(ArrayList<MiniSite> miniSites) {
+                        setMiniSites(miniSites);
+                    }
+                }, mParentActivity.getUserId());
+        mNetworkManager.getRequestQueue().add(miniSitesRequest);
+    }
+
     public abstract void refreshTaskList(List<Task> tasks);
 
     public void setSwipeFalse() {
@@ -182,5 +203,12 @@ public abstract class TaskListAbstractFragment extends ListFragment {
         mTaskList.clear();
         mTaskList.putAll(tasks);
         mTaskAdapter.notifyDataSetChanged();
+    }
+
+    private void setMiniSites(ArrayList<MiniSite> miniSites){
+        mUserMiniSiteList.clear();
+        for (MiniSite miniSite : miniSites){
+            mUserMiniSiteList.add(miniSite);
+        }
     }
 }
